server-python/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
mysql-connector-python==9.2.0
namex==0.0.8
numba==0.61.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pickle-mixin==1.0.2
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
pycparser==2.22
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
